+++ albaing-api/src/main/java/com/jobjob/albaing/controller/CompanyController.java
@@ -12,12 +12,5 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/account/auth")
 public class CompanyController {
 
-    @Autowired
-    private CompanyServiceImpl companyService;
 
-    // 기업 회원가입
-    @PostMapping("/register-company")
-    public void registerCompany(@RequestBody Company company) {
-        companyService.registerCompany(company);
-    }
 }
